requirements.txt
python-docx 
PyPDF2 
chromadb
sentence_transformers
nltk
PyMuPDF
pdfplumber
fastapi
python-multipart
python-dotenv
uvicorn
pymongo